% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagerankr-package.R
\docType{package}
\name{pagerankr-package}
\alias{pagerankr-package}
\alias{pagerankr}
\title{pagerankr: A Modular Toolkit for PageRank Calculation}
\description{
The pagerankr package provides a set of modular, pipeable functions 
to calculate PageRank scores from edge lists and redirect reports, 
common in SEO analysis. Functions handle URL cleaning, redirect resolution, 
edge deduplication, isolate handling, and PageRank computation using base R 
for data manipulation and 'igraph' for core PageRank calculation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bart-turczynski/pagerankr}
  \item Report bugs at \url{https://github.com/bart-turczynski/pagerankr/issues}
}

}
\author{
\strong{Maintainer}: Bart Turczynski \email{bartek@turczynski.pl}

}
\keyword{internal}
