% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screaming_frog_bundle.R
\name{print.screaming_frog_bundle}
\alias{print.screaming_frog_bundle}
\title{Print a Screaming Frog bundle}
\usage{
\method{print}{screaming_frog_bundle}(x, ...)
}
\arguments{
\item{x}{A `screaming_frog_bundle` object.}

\item{...}{Unused; for S3 compatibility.}
}
\value{
`x`, invisibly.
}
\description{
Print a Screaming Frog bundle
}
