% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_audit.R
\name{print.transition_audit}
\alias{print.transition_audit}
\title{Print a transition_audit object}
\usage{
\method{print}{transition_audit}(x, ...)
}
\arguments{
\item{x}{A `transition_audit` object.}

\item{...}{Unused; for S3 compatibility.}
}
\value{
`x`, invisibly.
}
\description{
Print a transition_audit object
}
